Rename testing_resources.py to resources.py
